EHS Botball 2022 Program with Changes
We added a while loop so the bot moves forward when it senses no red or green objects (no objects in channels 0 or 1).  This causes the bot to be able to move up to the pom so it is in range of the sweeper arm on the servo to hit the colored pom to the correct direction for sorting.